fix(client): report the seek target upstream instead of the pre-seek position
Follow-up to the warm-pause in-place seek fix. On a SYNCED display the
seeked player no longer re-resolves, but every seek still landed back at
the pre-seek spot: doSeek() fired events.onSeek() right after submitting
the asynchronous libvlc set_time, and DisplayScreen.afterSeek() then read
the live player position for the PlaybackCommand - which at that instant
still reported the pre-seek spot. The server seeked its authoritative
timeline to that old position, and the follower dutifully dragged every
viewer back to it (the visible pause-then-resume was the corrective
in-place seek freezing the picture).

The onSeek event and PlaybackHost.afterSeek() now carry the requested
target nanos, and DisplayScreen emits the SEEK intent with that target.
A loop restart (restartFromBeginning) reports 0.

+++ media/player/src/main/kotlin/com/dreamdisplayx/media/player/events/PlayerEvents.kt
@@ -10,6 +10,6 @@ internal data class PlayerEvents(
     /** Called on an unrecoverable playback error. */
     val onError: (DreamMediaException) -> Unit,
 
-    /** Called after a seek completes so the screen can reset its overlay state. */
-    val onSeek: () -> Unit,
+    /** Called after a seek with the requested target (nanos) so the screen can reset its overlay state. */
+    val onSeek: (Long) -> Unit,
 )
